backend/requirements.txt
# CodeRecall Backend Dependencies

# FastAPI
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5

# Database
sqlalchemy==2.0.46
alembic==1.18.4
psycopg2-binary==2.9.11

# Authentication
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.21

# Environment
python-dotenv==1.2.1

# Production WSGI server
gunicorn==23.0.0

# Rate limiting
slowapi==0.1.9

# Email Service (SendGrid)
sendgrid==6.12.5

# Scheduler for automated emails
APScheduler==3.11.2
pytz==2025.2

# Email validation
email-validator==2.3.0

# Google OAuth
google-auth==2.38.0
requests==2.32.3
